BigQueryIO: move to GCP module and package
